third_party/blink/web_tests/wpt_internal/css/css-conditional/container-queries/scroll-state/scroll-direction-user-touch-scroll-expected.txt

This is a testharness.js-based test.
[FAIL] Fling gesture scroll is a relative scroll
  assert_equals: expected "yes" but got "no"
Harness: the test ran to completion.
